pyproject.toml
[project]
name = "code2sql"
version = "0.1.0"
description = "Code to SQL generation tool with LLM support"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.0.0",
    "httpx>=0.24.0",
    "aiohttp>=3.8.0",
    "PyYAML>=6.0",
    "sqlglot>=20.0.0",
    "tqdm>=4.65.0",
    "pandas>=2.0.0",
    "pyarrow>=20.0.0",
    "fastparquet>=2024.11.0",
    # Web Server Dependencies
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "jinja2",
    # ML / Transformers Dependencies
    "torch",
    "transformers",
    "accelerate",
    "bitsandbytes",
    "openpyxl>=3.1.5",
    "matplotlib>=3.10.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["config", "data_processing", "model", "sql_generation", "utils", "workflows"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
